# openapi.json
{
  "info": {
    "title": "Getting started - Backend",
    "version": "1.0.0",
    "description": "Bem vindos a iniciativa: Getting started - Backend! :)"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/pessoas/": {
      "get": {
        "operationId": "PessoaApiSpec_get_/pessoas/",
        "tags": [
          "Pessoa"
        ],
        "summary": "Lista de pessoas",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pessoas": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pessoa"
                      }
                    }
                  },
                  "additionalProperties": false,
                  "required": [
                    "pessoas"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "PessoaApiSpec_post_/pessoas/",
        "tags": [
          "Pessoa"
        ],
        "summary": "Cria uma Pessoa",
        "parameters": [],
        "requestBody": {
          "description": "Pessoa do treinamento",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "Pessoa do treinamento",
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "altura": {
                    "type": "number"
                  },
                  "dataNascimento": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "altura",
                  "dataNascimento",
                  "id",
                  "nome"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pessoa do treinamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      }
    },
    "/api/pessoas/{id}": {
      "get": {
        "operationId": "PessoaApiSpec_get_/pessoas/{id}",
        "tags": [
          "Pessoa"
        ],
        "summary": "Obtem Pessoa pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pessoa do treinamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "PessoaApiSpec_put_/pessoas/{id}",
        "tags": [
          "Pessoa"
        ],
        "summary": "Atualiza uma Pessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Pessoa do treinamento",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "Pessoa do treinamento",
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "nome": {
                    "type": "string"
                  },
                  "altura": {
                    "type": "number"
                  },
                  "dataNascimento": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "additionalProperties": false,
                "required": [
                  "altura",
                  "dataNascimento",
                  "id",
                  "nome"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pessoa do treinamento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pessoa"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "PessoaApiSpec_delete_/pessoas/{id}",
        "tags": [
          "Pessoa"
        ],
        "summary": "Remove uma Pessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "components": {
    "schemas": {
      "Pessoa": {
        "description": "Pessoa do treinamento",
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nome": {
            "type": "string"
          },
          "altura": {
            "type": "number"
          },
          "dataNascimento": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "required": [
          "altura",
          "dataNascimento",
          "id",
          "nome"
        ]
      }
    }
  }
}